Fix adding multiple references into a single variant + allow 'ordering'
Note: The ordering on publish for this is currently mostly unused since they layer describes the ordering, and the variant references would only overlap if they were to be prepended in the same layer to the same variant. As such, currently they will always have the same ordering - which means the new publish being added (even if being replaced) will currently end up last in the prepended items, and thus be weaker than the other layers. In short, the layer order might behave 'unexpectedly' currently.